+++ gpt4all-backend/src/llmodel_shared.cpp
@@ -161,7 +161,9 @@ bool LLModel::decodePrompt(std::function<bool(int32_t)> promptCallback,
                            std::vector<Token> embd_inp,
                            bool isResponse) {
     if ((int) embd_inp.size() > promptCtx.n_ctx - 4) {
-        responseCallback(-1, "ERROR: The prompt size exceeds the context window size and cannot be processed.");
+        // FIXME: (Adam) We should find a way to bubble these strings to the UI level to allow for
+        // translation
+        responseCallback(-1, "Your message was too long and could not be processed. Please try again with something shorter.");
         std::cerr << implementation().modelType() << " ERROR: The prompt is " << embd_inp.size() <<
             " tokens and the context window is " << promptCtx.n_ctx << "!\n";
         return false;
@@ -244,6 +246,8 @@ void LLModel::generateResponse(std::function<bool(int32_t, const std::string&)> 
         return;
     }
 
+    initSampler(promptCtx);
+
     std::string cachedResponse;
     std::vector<Token> cachedTokens;
     int n_predicted = 0;
@@ -251,7 +255,7 @@ void LLModel::generateResponse(std::function<bool(int32_t, const std::string&)> 
     // Predict next tokens
     for (bool stop = false; !stop;) {
         // Sample next token
-        std::optional<Token> new_tok = sampleToken(promptCtx);
+        std::optional<Token> new_tok = sampleToken();
         std::string new_piece = tokenToString(new_tok.value());
         cachedTokens.push_back(new_tok.value());
         cachedResponse += new_piece;