+++ gpt4all-chat/src/server.h
@@ -2,6 +2,7 @@
 #define SERVER_H
 
 #include "chatllm.h"
+#include "chatmodel.h"
 #include "database.h"
 
 #include <QHttpServer>
@@ -32,7 +33,7 @@ public Q_SLOTS:
     void start();
 
 Q_SIGNALS:
-    void requestServerNewPromptResponsePair(const QString &prompt);
+    void requestServerNewPromptResponsePair(const QString &prompt, const QList<PromptAttachment> &attachments = {});
 
 private:
     auto handleCompletionRequest(const CompletionRequest &request) -> std::pair<QHttpServerResponse, std::optional<QJsonObject>>;